refactor(frontend): extract dashboard viewer filter
